// src/app/models/admin/map/layer.model.ts
import { DeserializableModel } from "src/app/models/admin/deserializable.model";

export class OutputLayerModel implements DeserializableModel {
  id: number;
  layerType: string;
  projectionId: number;
  layerName: string;
  bingKey: string;
  bingImageryset: string;
  mapquestLayer: string;
  wmsServertype: string;
  wmsUrl: string;
  wmsInfoUrl: string;
  wmsLayers: string;
  wmsTiled: boolean;
  wmsVersion: string;
  wfsUrl: string;
  wfsTypename: string;
  featureInfoTemplate: string;
  wfsVersion: string;
  gpxFilename: string;
  kmlFilename: string;
  postgisSchema: string;
  postgisHost: string;
  postgisPort: string;
  postgisUser: string;
  postgisPass: string;
  postgisDbname: string;
  postgisTable: string;
  postgisField: string;
  postgisAttributes: string;
  geojsonGeomtype: string;
  geojsonAttributes: string;
  geojsonFeatures: string;
  shapefileFilename: string;
  shapefileGeomtype: string;
  shapefileWmsurl: string;
  shapefileMsclass: string;
  geopackageFilename: string;
  geopackageTable: string;
  geopackageFields: string;
  hasz: string;
  copyright: string;
  moreSettings: string;
  description: string;
  layerTitle: string;
  opacity: number;
  status: number;
  extentMinx: number;
  extentMiny: number;
  extentMaxx: number;
  extentMaxy: number;
  minResolution: number;
  maxResolution: number;
  fieldsInfo: string;
  fieldsDisplay: string;
  fieldsInfoFormat: string;
  fieldsInfoTemplate: string;
  deserialize(input: any): this {
    return Object.assign(this, input);
  }
}

export class InputLayerModel {
  id: number;
  layerType: string;
  projectionId: number;
  layerName: string;
  bingKey: string;
  bingImageryset: string;
  mapquestLayer: string;
  wmsServertype: string;
  wmsUrl: string;
  wmsInfoUrl: string;
  wmsLayers: string;
  wmsTiled: boolean;
  wmsVersion: string;
  wfsUrl: string;
  wfsTypename: string;
  featureInfoTemplate: string;
  wfsVersion: string;
  gpxFilename: string;
  kmlFilename: string;
  postgisSchema: string;
  postgisHost: string;
  postgisPort: string;
  postgisUser: string;
  postgisPass: string;
  postgisDbname: string;
  postgisTable: string;
  postgisField: string;
  postgisAttributes: string;
  geojsonGeomtype: string;
  geojsonAttributes: string;
  geojsonFeatures: string;
  shapefileFilename: string;
  shapefileGeomtype: string;
  shapefileWmsurl: string;
  shapefileMsclass: string;
  geopackageFilename: string;
  geopackageTable: string;
  geopackageFields: string;
  hasz: string;
  copyright: string;
  moreSettings: string;
  description: string;
  layerTitle: string;
  opacity: number;
  status: number;
  extentMinx: number;
  extentMiny: number;
  extentMaxx: number;
  extentMaxy: number;
  minResolution: number;
  maxResolution: number;
  fieldsInfo: string;
  fieldsDisplay: string;
  fieldsInfoFormat: string;
  fieldsInfoTemplate: string;
}