// src/app/shared/constants/consts/enum.ts
export enum loaigiayto {
  Cmnd = 'Chứng minh nhân dân',
  Hc = 'Hộ chiếu',
}
export enum donvido {
  m = 0,
  cm,
  mol,
  l,
}

export enum ValidateMaxLength {
  longText = 255,
  shortLongText = 150,
  mediumText = 40,
  shortText = 10
}

// export enum ValidateMaxLength {
//   longText = 5,
//   shortLongText = 4,
//   mediumText = 3,
//   shortText = 2
// }